requirements.txt
pymongo
pydantic
Flask
python-dotenv
gunicorn
bcrypt
PyJWT
flask_cors
pillow-heif